Module refactor part 2
