+++ gneiss-mqtt/src/mqtt/suback.rs
@@ -5,13 +5,11 @@
 
 use crate::decode::*;
 use crate::encode::*;
-use crate::encode::utils::*;
 use crate::error::{MqttError, MqttResult};
 use crate::logging::*;
 use crate::mqtt::*;
 use crate::mqtt::utils::*;
 use crate::validate::*;
-use crate::validate::utils::*;
 
 use std::collections::VecDeque;
 use std::fmt;
@@ -276,7 +274,6 @@ mod tests {
     }
 
     use crate::validate::testing::*;
-    use crate::validate::utils::testing::*;
 
     test_ack_validate_failure_inbound_packet_id_zero!(suback_validate_failure_internal_packet_id_zero, Suback, create_suback_all_properties, PacketType::Suback);
 }
