+++ gneiss-mqtt/src/mqtt/subscribe.rs
@@ -6,13 +6,11 @@
 #[cfg(test)]
 use crate::decode::*;
 use crate::encode::*;
-use crate::encode::utils::*;
 use crate::error::{MqttError, MqttResult};
 use crate::logging::*;
 use crate::mqtt::*;
 use crate::mqtt::utils::*;
 use crate::validate::*;
-use crate::validate::utils::*;
 
 use std::collections::VecDeque;
 use std::fmt;
@@ -249,7 +247,7 @@ mod tests {
 
     use super::*;
     use crate::decode::testing::*;
-    use crate::validate::utils::testing::verify_validation_failure;
+    use crate::validate::testing::*;
 
     #[test]
     fn subscribe_round_trip_encode_decode_default() {
@@ -379,8 +377,6 @@ mod tests {
 
         do_inbound_size_decode_failure_test(&MqttPacket::Subscribe(packet));
     }
-
-    use crate::validate::testing::*;
 
     #[test]
     fn unsubscribe_validate_success() {
