+++ gneiss-mqtt/src/mqtt/unsuback.rs
@@ -5,13 +5,11 @@
 
 use crate::decode::*;
 use crate::encode::*;
-use crate::encode::utils::*;
 use crate::error::{MqttError, MqttResult};
 use crate::logging::*;
 use crate::mqtt::*;
 use crate::mqtt::utils::*;
 use crate::validate::*;
-use crate::validate::utils::*;
 
 use std::collections::VecDeque;
 use std::fmt;
@@ -278,7 +276,6 @@ mod tests {
     }
 
     use crate::validate::testing::*;
-    use crate::validate::utils::testing::*;
 
     test_ack_validate_failure_inbound_packet_id_zero!(unsuback_validate_failure_internal_packet_id_zero, Unsuback, create_unsuback_all_properties, PacketType::Unsuback);
 }
