+++ gneiss-mqtt/src/protocol.rs
@@ -10,7 +10,6 @@ use crate::client::*;
 use crate::config::*;
 use crate::decode::*;
 use crate::encode::*;
-use crate::encode::utils::*;
 use crate::error::{fold_mqtt_result, MqttError, MqttResult};
 use crate::mqtt::*;
 use crate::mqtt::connack::*;
